task/requirements.txt
# requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
requests
psycopg2-binary
python-dotenv
